src/test/javascript/spec/app/entities/m-pvp-ranking-reward-group/m-pvp-ranking-reward-group.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { MPvpRankingRewardGroupService } from 'app/entities/m-pvp-ranking-reward-group/m-pvp-ranking-reward-group.service';
import { IMPvpRankingRewardGroup, MPvpRankingRewardGroup } from 'app/shared/model/m-pvp-ranking-reward-group.model';

describe('Service Tests', () => {
  describe('MPvpRankingRewardGroup Service', () => {
    let injector: TestBed;
    let service: MPvpRankingRewardGroupService;
    let httpMock: HttpTestingController;
    let elemDefault: IMPvpRankingRewardGroup;
    let expectedResult;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = {};
      injector = getTestBed();
      service = injector.get(MPvpRankingRewardGroupService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new MPvpRankingRewardGroup(0, 0, 0, 0, 0);
    });

    describe('Service methods', () => {
      it('should find an element', async () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: elemDefault });
      });

      it('should create a MPvpRankingRewardGroup', async () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new MPvpRankingRewardGroup(null))
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should update a MPvpRankingRewardGroup', async () => {
        const returnedFromService = Object.assign(
          {
            groupId: 1,
            contentType: 1,
            contentId: 1,
            contentAmount: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should return a list of MPvpRankingRewardGroup', async () => {
        const returnedFromService = Object.assign(
          {
            groupId: 1,
            contentType: 1,
            contentId: 1,
            contentAmount: 1
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a MPvpRankingRewardGroup', async () => {
        const rxPromise = service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
